+++ cablecolor.h
@@ -1,0 +1,10 @@
+#ifndef CABLECOLOR_H
+#define CABLECOLOR_H
+
+namespace TelCoColorCoder
+{
+    enum MajorColor { WHITE, RED, BLACK, YELLOW, VIOLET };
+    enum MinorColor { BLUE, ORANGE, GREEN, BROWN, SLATE };
+}
+
+#endif
