+++ colorpair.cpp
@@ -1,0 +1,20 @@
+#include <string.h>
+
+#include "colorpair.h"
+
+namespace TelCoColorCoder {
+     
+    MajorColor ColorPair::getMajor() {
+        return majorColor;
+    }
+    MinorColor ColorPair::getMinor() {
+        return minorColor;
+    }
+    std::string ColorPair::ToString() {
+        std::string colorPairStr = MajorColorNames[majorColor];
+        colorPairStr += " ";
+        colorPairStr += MinorColorNames[minorColor];
+        return colorPairStr;
+    }
+}
+
